Move core logic to package
