+++ curve_dao/vote_utils.py
@@ -1,11 +1,17 @@
 import json
 import os
+import sys
 import warnings
 from typing import Dict, List, Tuple
 
 import ape
 import requests
 from ape.logging import logger
+
+from curve_dao.decoder_utils import decode_input
+
+warnings.filterwarnings("ignore")
+
 
 warnings.filterwarnings("ignore")
 
@@ -90,3 +96,84 @@ def make_vote(target: Dict, actions: List[Tuple], description: str, vote_creator
     )
 
     return tx
+
+
+def decode_vote_script(script):
+    idx = 4
+
+    votes = []
+    while idx < len(script):
+
+        # get target contract address:
+        target = ape.Contract(script[idx : idx + 20])
+        idx += 20
+
+        length = int(script[idx : idx + 4].hex(), 16)
+        idx += 4
+
+        # calldata to execute for the dao:
+        calldata = script[idx : idx + length]
+        idx += length
+
+        fn, inputs = decode_input(target, calldata)
+        agent = None
+
+        # print decoded vote:
+        if calldata[:4].hex() == "0xb61d27f6":
+            agent = target
+            target = ape.Contract(inputs[0])
+            fn, inputs = decode_input(target, inputs[2])
+            inputs_with_names = get_inputs_with_names(fn, inputs)
+            formatted_inputs = format_fn_inputs(inputs_with_names)
+            formatted_output = (
+                f"Call via agent: [yellow]{agent}[/]\n"
+                f" ├─ [bold]To[/]: [green]{target}[/]\n"
+                f" ├─ [bold]Function[/]: [yellow]{fn.name}[/]\n"
+                f" └─ [bold]Inputs[/]: \n{formatted_inputs}\n"
+            )
+        else:
+            inputs_with_names = get_inputs_with_names(fn, inputs)
+            formatted_inputs = format_fn_inputs(inputs_with_names)
+            formatted_output = (
+                f"Direct call\n "
+                f" ├─ [bold]To[/]: [green]{target}[/]\n"
+                f" ├─ [bold]Function[/]: [yellow]{fn.name}[/]\n"
+                f" └─ [bold]Inputs[/]: {formatted_inputs}\n"
+            )
+
+        vote = {
+            "agent": agent.address if agent else None,
+            "target": target.address,
+            "function": fn.name,
+            "inputs": inputs_with_names,
+            "formatted_output": formatted_output,
+        }
+        votes.append(vote)
+
+    return votes
+
+
+def get_inputs_with_names(abi, inputs):
+    arg_names = []
+    for i in range(len(inputs)):
+        argname = abi.inputs[i].name
+        arg_names.append(argname)
+
+    inputs_with_names = list(zip(arg_names, inputs))
+    return inputs_with_names
+
+
+def format_fn_inputs(inputs_with_names):
+    if len(inputs_with_names) == 0:
+        return ""
+
+    if len(inputs_with_names) == 1:
+        name, arg = inputs_with_names[0]
+        return f"    └─ [bold]{name}[/]: [yellow]{arg}[/]"
+
+    formatted_args = ""
+    for name, arg in inputs_with_names[:-1]:
+        formatted_args += f"    ├─ [bold]{name}[/]: [yellow]{arg}[/]\n"
+    name, arg = inputs_with_names[-1]
+    formatted_args += f"    └─ [bold]{name}[/]: [yellow]{arg}[/]"
+    return formatted_args
